+++ src/Commands.cpp
@@ -316,6 +316,7 @@ static SeqStrings gCommandNames =
     "CmdShowAnnotationText\0"
     "CmdAnnotationHighlightBrush\0"
     "CmdFindAnnotation\0"
+    "CmdOpenFileNoHistory\0"
     "CmdNone\0"
     "\0";
 
@@ -620,6 +621,7 @@ static i32 gCommandIds[] = {
     CmdShowAnnotationText,
     CmdAnnotationHighlightBrush,
     CmdFindAnnotation,
+    CmdOpenFileNoHistory,
     CmdNone,
 };
 
@@ -924,6 +926,7 @@ SeqStrings gCommandDescriptions =
     "Show Comment\0"
     "Highlight with Brush\0"
     "Find Annotation\0"
+    "Open File Without History...\0"
     "Do nothing\0"
     "\0";
 // clang-format on